scripts: support for inserting clocks into serial devs
Previously, serial devices used the property "clock-frequency".
With this commit, support is given to supplant the "clock-frequency"
property with an actual "clocks" property, and to attach the
appropriate UART clock to the serial device.

Note that the stock Linux 8250 (nsl6550a) can take either
a "clock-frequency" or a "clocks" property.  So it is
okay if a Bolt image with this commit is run on a Linux
without the new uart driver.  If the new uart driver is
present, the driver takes advantage of the "clocks" property
by using the MUX behind it; if the new driver only sees
the "clock-frequency" property, it defaults to ns16550a
behavior.

Any new commmits with clock files (clks-XXXXXX.plx) where
UART functions are defined MUST be ack'd after this commit.
If ack'd before, additional clocks will be defined and the
"clock-frequency" prop will be used for the ns16550a,
resulting in the turning off of the clocks by Linux
since there was no driver to claim them.

refs #SWBOLT-2207

Signed-off-by: Jim Quinlan <james.quinlan@broadcom.com>

Conflicts:
	scripts/BcmDt/Devices.pm