android: remove "AB_VERITY_PUB_KEY" logic
This commit removes the "AB_VERITY_PUB_KEY" logic and lets boot_control mask
slot success.

With this change, the bootloader is not compatible with N (or earlier)
devices which do not implement dm-verity enforcing (i.e. development setup
for the most part). This is proper because it is not expected that any N
devices will require to update bootloader unless they are also updating to
O for full 'treble' support.

refs #SWANDROID-4078

Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>
[amended commit message]
Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>