android: dm-verity corruption detection support.
if bootloader reset reason shows that dm-verity partition was
corrupted, switch dm-verity boot mode from enforcing to eio to
allow android to still attempt booting (per expectation) leaving
out the corrupted blocks aside.

once we detect such corruption, the only way to clear it is to either
erase or flash a new partition of the slot set, in which case we grant
a clean start again (and if corruption persists, it will be marked again
anyways).

refs #SWANDROID-4283

Change-Id: I927d341ed34d3aae01d01a432479ca530dcc6d64
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>