shmoo: v5.4.2.0 for 7268b0 and 7271b0
v5.4.2.0 is the M1 Shmoo release for 7268b0 and 7271b0. It contains
stability improvements for reliable 3733Mbps operation. It is
recommended to update to v5.4.2.0. Please note that Samsung parts
are not supported by this release.

As the API has changed, dropping the header and library into
existing locatins (replacing them with new ones) is not enough.
The API change has to be reflected into the make file.

Please also note that MCB's have to be regenerated.

SHA1(shmoo/B220_H00/lib/memsysinitlib.a)=
    e345d3060a81aa1ecc2e58b7916657830120f9dc

refs #SWBOLT-2438

Signed-off-by: Shane Yi <shane.yi@broadcom.com>