android: allow support for legacy layout with verity.
to keep the bootloader compatible with all the various usages,
we have to inject the verity knowledge as part of the build for
the device since that is the only way to distinguish a legacy
device with or without verity enabled.

adding a build time define since the bootloader is always device
specific, not the most elegant but certainly the less invasive,
as older device are retired in the future, this can also be removed.

refs #SWANDROID-4538
b/68717936

Change-Id: If770c5b9e2ba9686c5a3d085af2dec60d97f021e
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>