android: pass android_blkdev to kernel cmdline
In legacy boot mode, pass the UUID of a partition on the Android
block device to the kernel. This allows the kernel to wait for
the device to be available before launching init. This is needed
for the treble early mount feature to work consistently on legacy
boot platforms. The logic in Bolt and Kernel assumes that all
partitions that need early mounting exist on a since block device,
which is true at the moment. Will need to evolve if this changes.

refs #SWANROID-4568

Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>