script: fix filename arg used in 'do' operation
Applied the official fix to an earlier change on fixing builds on 17.04
ubuntu.

Revert "Fix avko and elfin builds on 17.04 ubuntu, debian stable and go/rodete."

This reverts commit a5b609266419e6e31ade0f3c06d3409c30bff390.

scripts: BcmDt: Provide a "clocks" property to the wake-timer

Provide a "clocks" property to the wake-timer which points to the 27Mhz
fixed-clock "upg_fixed" since this is the clock feed.

This is necessary for us to be able to use the Linux common clock
framework and remove the assumption about 27Mhz being the base frequency
in the driver.

refs #SWBOLT-2369

Signed-off-by: Florian Fainelli <florian.fainelli@broadcom.com>
Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>

Conflicts:
	scripts/config.pl

(cherry picked from commit 7ef5823b0f16f8711e21e9c3b3fe9d8e37ca214c)

script: fix filename arg used in 'do' operation

We use the 'do' operator to read in the clock configuration files.
Newer versions of Perls specify if a filename does not have a full
path or a preceding './' or '../', the filename will be searched for
in @INC only.  This is not what we want so we prepend our clock file
name with './'.

In addition, two functions that were using the filename have been
changed so that they only use a bool to denote its existence.

refs #SWBOLT-2397

Signed-off-by: Jim Quinlan <james.quinlan@broadcom.com>
Signed-off-by: Danesh Petigara <danesh.petigara@broadcom.com>

Conflicts:
	scripts/config.pl

(cherry picked from commit ff2c16c4cad6360546320f77a2e34d98cb18b854)

Change-Id: Iba8702ca2436416b61c7eded979e2fc9abd6361e