shmoo: v5.4.9.0 for 7268b0 and 7271b0
SHA1(shmoo/B220_H00/lib/memsysinitlib.a)=
    53969e00f1622c84d493f5d8719354e22470c39c

changes comparing against v5.4.8.0:
- overflow issue that caused Write Leveling to fail on some boards
- adjusts Read to Write turnaround delay for increased stability

changes comparing against v5.4.2.0:
- Support for Samsung 14nm (single Rank) LPDDR4
- Temp Readout
- Dynamic Tuning
- stability improvement

v5.4.9.0 (starting v5.4.8.0) comes with a DPFE firmware (v1.1.0.0) and
loads the firmware while Shmoo'ing so that Linux does not have to do
it. Letting Linux load a firmware is not possible on secure parts
because the DPFE DMEM and IMEM are not accessible from the host
processor past a certain point of the boot process.

The DPFE firmware is responsible for tempertature reading and dynamic
tuning.

refs #SWBOLT-2541
refs #SWANDROID-4708

Signed-off-by: Shane Yi <shane.yi@broadcom.com>
Signed-off-by: Pierre Couillaud <pierre@broadcom.com>