android:bsu: update licensing to Apache 2.0.
all the BSU code is written by Broadcom, therefore there is no
need to have it licensed as GPL.  to clarify the ownership and
avoid confusion we re-license the whole BSU code as Apache 2.0
(which is the standard Android license); this allows customer to
also freely re-use whole or part of the BSU delivered by Broadcom
without concerns of GPL pollution.

we choose Apache 2.0 license because there is nothing proprietary
in the BSU code and we may at some point incorporate some Android
content which is using such license as well.

the only GPL aspect of this code is the "part_efi.h" header which has
been cleaned out of any potential copyrightable information to allow
usage in the code (the original GPL licensed header from which the
cleaned up version is derived has been preserved for reference under
part_efi.h.original).

refs #SWANDROID-4764

Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>