shmoo: v5.6.1.0 for 7268b0 and 7271b0
SHA1(shmoo/B220_H00/lib/memsysinitlib.a)=
    1cd259db37eeef02ac50933788dd476b4fc955a1

The previous memsysinitlib version 5.6.0.0 has accidentally disabled
the dynamic tuning. The dynamic tuning is required to respond to the
on-going temperature of a system (specifically LPDDR4 DRAM) if it is
quite different than the temperature of when Shmoo is done.
Otherwise, the system might be instable.

The memsysinitlib v5.6.0.0 has to be updated to v5.6.1.0 as quickly
as possible.

Since the default toolchain has not changed to 6.3-1.2, the
memsysinitlib that had been built by 4.8 has been integrated.

refs #SWBOLT-2632

Signed-off-by: Shane Yi <shane.yi@broadcom.com>
(cherry picked from commit 9eff10a14b374357f2188e98f9659ab9eaf9bc3d)